support agent/rag-support-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.5
numpy==1.24.3
scikit-learn==1.3.2
httpx==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
python-json-logger==2.0.7
sentence-transformers==2.2.2
rank-bm25==0.2.2
pdfplumber==0.10.3
python-docx==0.8.11
celery==5.3.4
aiofiles==23.2.1
